+++ src/libTAS/renderhud/RenderHUD_GL.h
@@ -33,6 +33,13 @@ class RenderHUD_GL : public RenderHUD
     public:
         RenderHUD_GL();
         ~RenderHUD_GL();
+
+        /* Initialize texture and fbo */
+        void init();
+
+        /* Deallocate texture and fbo */
+        void fini();
+
         void renderText(const char* text, Color fg_color, Color bg_color, int x, int y);
     private:
         static GLuint texture;